betterAI/requirements.txt
youtube_transcript_api
pathlib
google-generativeai 
streamlit 
python-dotenv
PyPDF2
faiss-cpu
langchain_google_genai
langchain